+++ dwds/CHANGELOG.md
@@ -8,6 +8,8 @@
   to enable running expression compiler worker from
   a given location.
 - Support Chrome `skipLists` to improve stepping performance.
+- Depend on the latest `package:vm_service` which supports a new
+  `limit` parameter to `getStack`.
 
 ## 7.0.2